cleanup and action masking
