+++ utils/obs/__init__.py
@@ -1,0 +1,6 @@
+class ObsGenerator:
+    def __init__(self):
+        pass
+
+    def calc_obs(self, obs):
+        raise NotImplementedError
