+++ utils/reward/__init__.py
@@ -1,0 +1,6 @@
+class RewardGenerator:
+    def __init__(self):
+        pass
+
+    def calc_rewards(self, old_obs, actions, obs):
+        raise NotImplementedError
